Move TOC after title page in pdf preview - Move document title to a separate component from document content
